requirements.txt
discord.py>=2.3.2
aiosqlite>=0.17.0
python-dotenv>=1.0.0